Updated view as per PSLab
